Make CameraAwareGroup layered
This is so that the player can be on top of all other actors. In the
future there may be other uses.

CameraAwareGroup is now CameraAwareLayeredGroup and is a subclass of
pygame.sprite.LayeredUpdates, but the draw() method is unchanged because
tracking updated regions of the screen isn't too useful when the
background is made of sprites and can be animated.