Add Actor.set_rect() method
Instead of just assigning a rect to the Actor's rect attribute, ignoring
its position, use a method to set the rect attribute but also update the
rect's position. If the rect's position is not set right away, the Actor
will test for collision with a Y-position of 0 on the first frame.